+ CMakeLists.txt

+++ blink/blink.c
@@ -3,24 +3,26 @@
 #include <stdlib.h>
 #include <unistd.h>
 
-int main (int argc,char* argv[])
+int main(int argc, char* argv[])
 {
-  if (argc < 2) {
-    printf("Usage example: ./blink gpio_port\n");
-    return 1;
-  }
-  int gpioPort = atoi(argv[1]);
+    if (argc < 2) 
+    {
+        printf("Usage example: ./blink gpio_port\n");
+        return 1;
+    }
+    int gpioPort = atoi(argv[1]);
 
-  wiringPiSetup();
+    wiringPiSetup();
 
-  pinMode(gpioPort, OUTPUT);  // set mode to output
+    pinMode(gpioPort, OUTPUT);  // set mode to output
 
-  int level = 0;
-  while(1) {
-    level = (0 == level) ? 1 : 0;
-    digitalWrite(gpioPort, level); // output a low/high level to the specified GPIO port
-    sleep(1);
-  }
+    int level = 0;
+    while (1) 
+    {
+        level = (0 == level) ? 1 : 0;
+        digitalWrite(gpioPort, level); // output a low/high level to the specified GPIO port
+        sleep((unsigned int)1);
+    }
 
-  return 0;
+    return 0;
 }
